Ai model/requirements.txt
deepface
opencv-python 
mtcnn 
numpy 
pandas 
matplotlib
scikit-learn
tf-keras
fastapi
uvicorn
python-multipart
gradio
pillow
